Separated logging functionality from main SDK
To save memory in the Ardusat SDK, the functionality to log to a SD card
has been removed from the main SDK and moved into a separate library.

This library can be found at
https://github.com/ArduSat/ArdusatSDK-Logging.

Sketches that use the basic sensor `read` functions and `toCSV` or
`toJSON` functions will not need to change, but sketches that relied on
RTC functionality, or logging to a SD card will need to include the SDK
logging library separately.